+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Bugfixes
 
+- [\#214](https://github.com/arkworks-rs/r1cs-std/pull/214) Enforce `FpVar` conditional equality for unequal constants
 - [\#156](https://github.com/arkworks-rs/r1cs-std/pull/156) Fix panic in `impl Sum for FpVar`
 - [\#198](https://github.com/arkworks-rs/r1cs-std/pull/198) Remove duplicate native cases in run_binary_exhaustive_both